Bug Fix, convert int to float, integer division truncates
Problem with convert from int to float ->Method "CalculateAvgSpeed"
Problem with "integer division" ->Method "CalculateDistance"

+++ app/src/main/java/hr/foi/air1719/personaltracker/Helper.java
@@ -51,20 +51,20 @@ public class Helper {
         return false;
     }
 
-    public static double CalculateDistance(Location startPoint, Location endPoint)
+    public static float CalculateDistance(Location startPoint, Location endPoint)
     {
-        return (startPoint.distanceTo(endPoint)/1000);
+        return (startPoint.distanceTo(endPoint) / ((float)1000));
     }
 
-    public static double CalculateAvgSpeed(Date startDate, Date endDate, double KM)
+    public static float CalculateAvgSpeed(Date startDate, Date endDate, double KM)
     {
         try {
             long different = endDate.getTime() - startDate.getTime();
 
-            double diffHours = (different / (60 * 60 * 1000));
+            float diffHours = (float)((float)different / (float)(60 * 60 * 1000));
             if(diffHours==0) return 0;
 
-            return (KM / diffHours);
+            return (float)(KM / diffHours);
         }
         catch (Exception E)
         {
